Remove the Alt flag from ToUnicodeEx call on Windows
This flag was added so that editors don't insert characters when
Alt+Arrow keys are pressed. This has the effect of breaking Alt code
characters so the flag has been removed. Moreover, the Alt+Arrow keys
behaviour seems to be fixed on Windows 11 (but is still present on
Windows 10). Anyway, removing this flag seems like the lesser of two
evils.