Simulate mouse wheel events regardless of settings on Windows
By default, SendInput reverses the direction of wheel events if the
mouse setting to reverse the direction is enabled. Since macOS and X11
don't do that, and the behavior should be aligned, mouse wheel event
simulation on Windows now checks whether the direction is reversed, and
if so, multiplies the rotation by -1 to nullify the reversal and always
simulate the same direction regardless of the setting.